requirements.txt
langchain
langchain-community
langchain-core
langchain-chroma
langchain-huggingface
langchain-ollama
python-dotenv
sentence-transformers
